Please make the version (and documention build) reproducible.
Whilst working on the Reproducible Builds effort [0], we noticed
that python-multipletau could not be built reproducibly.

This is because it uses the current date which gets embedded into
the documentation file if you are building from a tarball release
and not Git.

Patch attached that uses SOURCE_DATE_EPOCH [1]. This was originally
filed in Debian as #912957 [2].

 [0] https://reproducible-builds.org/
 [1] https://reproducible-builds.org/specs/source-date-epoch/
 [2] https://bugs.debian.org/912957